+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/bean/entity/HibernateFilterInfo.java
@@ -1,0 +1,134 @@
+package com.dwarfeng.judge.impl.bean.entity;
+
+import com.dwarfeng.datamark.bean.jpa.DatamarkEntityListener;
+import com.dwarfeng.datamark.bean.jpa.DatamarkField;
+import com.dwarfeng.judge.sdk.util.Constraints;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import javax.persistence.*;
+import java.util.Optional;
+
+@Entity
+@IdClass(HibernateLongIdKey.class)
+@Table(name = "tbl_filter_info")
+@EntityListeners(DatamarkEntityListener.class)
+public class HibernateFilterInfo implements Bean {
+
+    private static final long serialVersionUID = 4477628933708284149L;
+
+    // -----------------------------------------------------------主键-----------------------------------------------------------
+    @Id
+    @Column(name = "id", nullable = false, unique = true)
+    private Long longId;
+
+    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    @Column(name = "enabled", nullable = false)
+    private boolean enabled;
+
+    @Column(name = "type", length = Constraints.LENGTH_TYPE, nullable = false)
+    private String type;
+
+    @Column(name = "param", columnDefinition = "TEXT")
+    private String param;
+
+    @Column(name = "remark", length = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    // -----------------------------------------------------------审计-----------------------------------------------------------
+    @DatamarkField(handlerName = "filterDatamarkHandler")
+    @Column(
+            name = "created_datamark",
+            length = com.dwarfeng.datamark.util.Constraints.LENGTH_DATAMARK_VALUE,
+            updatable = false
+    )
+    private String createdDatamark;
+
+    @DatamarkField(handlerName = "filterDatamarkHandler")
+    @Column(
+            name = "modified_datamark",
+            length = com.dwarfeng.datamark.util.Constraints.LENGTH_DATAMARK_VALUE
+    )
+    private String modifiedDatamark;
+
+    public HibernateFilterInfo() {
+    }
+
+    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    public HibernateLongIdKey getKey() {
+        return Optional.ofNullable(longId).map(HibernateLongIdKey::new).orElse(null);
+    }
+
+    public void setKey(HibernateLongIdKey idKey) {
+        this.longId = Optional.ofNullable(idKey).map(HibernateLongIdKey::getLongId).orElse(null);
+    }
+
+    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    public Long getLongId() {
+        return longId;
+    }
+
+    public void setLongId(Long longId) {
+        this.longId = longId;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    public String getCreatedDatamark() {
+        return createdDatamark;
+    }
+
+    public void setCreatedDatamark(String createdDatamark) {
+        this.createdDatamark = createdDatamark;
+    }
+
+    public String getModifiedDatamark() {
+        return modifiedDatamark;
+    }
+
+    public void setModifiedDatamark(String modifiedDatamark) {
+        this.modifiedDatamark = modifiedDatamark;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "(" +
+                "longId = " + longId + ", " +
+                "enabled = " + enabled + ", " +
+                "type = " + type + ", " +
+                "param = " + param + ", " +
+                "remark = " + remark + ", " +
+                "createdDatamark = " + createdDatamark + ", " +
+                "modifiedDatamark = " + modifiedDatamark + ")";
+    }
+}