+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/cache/FilterInfoCacheImpl.java
@@ -1,0 +1,95 @@
+package com.dwarfeng.judge.impl.cache;
+
+import com.dwarfeng.judge.sdk.bean.entity.FastJsonFilterInfo;
+import com.dwarfeng.judge.stack.bean.entity.FilterInfo;
+import com.dwarfeng.judge.stack.cache.FilterInfoCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class FilterInfoCacheImpl implements FilterInfoCache {
+
+    private final RedisBatchBaseCache<LongIdKey, FilterInfo, FastJsonFilterInfo> batchBaseCache;
+
+    public FilterInfoCacheImpl(RedisBatchBaseCache<LongIdKey, FilterInfo, FastJsonFilterInfo> batchBaseCache) {
+        this.batchBaseCache = batchBaseCache;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(LongIdKey key) throws CacheException {
+        return batchBaseCache.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public FilterInfo get(LongIdKey key) throws CacheException {
+        return batchBaseCache.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void push(FilterInfo value, long timeout) throws CacheException {
+        batchBaseCache.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(LongIdKey key) throws CacheException {
+        batchBaseCache.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void clear() throws CacheException {
+        batchBaseCache.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<LongIdKey> keys) throws CacheException {
+        return batchBaseCache.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<LongIdKey> keys) throws CacheException {
+        return batchBaseCache.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<FilterInfo> batchGet(@SkipRecord List<LongIdKey> keys) throws CacheException {
+        return batchBaseCache.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<FilterInfo> entities, long timeout) throws CacheException {
+        batchBaseCache.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<LongIdKey> keys) throws CacheException {
+        batchBaseCache.batchDelete(keys);
+    }
+}