+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/bean/entity/FilterInfo.java
@@ -1,0 +1,91 @@
+package com.dwarfeng.judge.stack.bean.entity;
+
+import com.dwarfeng.subgrade.stack.bean.entity.Entity;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+
+/**
+ * 过滤器信息。
+ *
+ * @author wangyc
+ * @since 2.3.0
+ */
+public class FilterInfo implements Entity<LongIdKey> {
+
+    private static final long serialVersionUID = 5788286232017499280L;
+
+    private LongIdKey key;
+    private boolean enabled;
+    private String type;
+    private String param;
+    private String remark;
+
+    public FilterInfo() {
+    }
+
+    public FilterInfo(
+            LongIdKey key,
+            boolean enabled,
+            String type,
+            String param,
+            String remark
+    ) {
+        this.key = key;
+        this.enabled = enabled;
+        this.type = type;
+        this.param = param;
+        this.remark = remark;
+    }
+
+    @Override
+    public LongIdKey getKey() {
+        return key;
+    }
+
+    @Override
+    public void setKey(LongIdKey key) {
+        this.key = key;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return "FilterInfo{" +
+                "key=" + key +
+                ", enabled=" + enabled +
+                ", type=" + type + '\'' +
+                ", param=" + param + '\'' +
+                ", remark=" + remark + '\'' +
+                '}';
+    }
+}